+++ apps/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @arche-cms/api
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @arche-cms/permissions@0.1.4
+
 ## 0.0.5
 
 ### Patch Changes